add sasin note
