+++ bots/telegram/sasin_char14.json
@@ -6,6 +6,7 @@
         "skill2": true,
         "skill3": true,
         "alive": true,
+        "noteObj": "",
         "life_cnt": 1,
         "whisper": 20,
         "note": 5,
@@ -24,6 +25,7 @@
         "skill2": true,
         "skill3": true,
         "alive": true,
+        "noteObj": "",
         "whisper": 20,
         "note": 5,
         "kill": 0,
@@ -41,6 +43,7 @@
         "skill2": true,
         "skill3": true,
         "alive": true,
+        "noteObj": "",
         "whisper": 20,
         "note": 5,
         "kill": 0,
@@ -58,6 +61,7 @@
         "skill2": true,
         "skill3": true,
         "alive": true,
+        "noteObj": "",
         "whisper": 20,
         "note": 5,
         "kill": 0,
@@ -75,6 +79,7 @@
         "skill2": true,
         "skill3": true,
         "alive": true,
+        "noteObj": "",
         "whisper": 20,
         "note": 5,
         "kill": 0,
@@ -92,6 +97,7 @@
         "skill2": true,
         "skill3": true,
         "alive": true,
+        "noteObj": "",
         "whisper": 20,
         "note": 5,
         "kill": 0,
@@ -110,6 +116,7 @@
         "skill3": true,
         "skill3_check": false,
         "alive": true,
+        "noteObj": "",
         "whisper": 20,
         "note": 5,
         "kill": 0,
@@ -127,6 +134,7 @@
         "skill2": true,
         "skill3": true,
         "alive": true,
+        "noteObj": "",
         "whisper": 20,
         "note": 5,
         "kill": 0,
@@ -145,6 +153,7 @@
         "skill2": true,
         "skill3": true,
         "alive": true,
+        "noteObj": "",
         "whisper": 20,
         "note": 5,
         "kill": 0,
@@ -163,6 +172,7 @@
         "skill3": true,
         "skill3_check": false,
         "alive": true,
+        "noteObj": "",
         "whisper": 20,
         "note": 5,
         "kill": 0,
@@ -180,6 +190,7 @@
         "skill2": true,
         "skill3": true,
         "alive": true,
+        "noteObj": "",
         "whisper": 20,
         "note": 5,
         "kill": 0,
@@ -198,6 +209,7 @@
         "skill3": true,
         "skill3_check": false,
         "alive": true,
+        "noteObj": "",
         "whisper": 20,
         "note": 5,
         "kill": 0,
@@ -215,6 +227,7 @@
         "skill2": true,
         "skill3": true,
         "alive": true,
+        "noteObj": "",
         "whisper": 20,
         "note": 5,
         "kill": 0,
@@ -232,6 +245,7 @@
         "skill2": true,
         "skill3": true,
         "alive": true,
+        "noteObj": "",
         "whisper": 20,
         "note": 5,
         "kill": 0,
